==> analysis/01_cleaning-data.rmd <==
---
title: "Cleaning of the data"
author: "Jono Tuke"
date: "Dec 2019"
output: 
  html_document: 
    number_sections: yes
    theme: spacelab
    toc: yes
    toc_float: yes
    df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE, 
  fig.width = 6, 
  fig.asp = 0.618, 
  out.width = "70%",
  fig.align = "center", 
  root.dir = '../'
)
pacman::p_load(tidyverse, readxl, 
               tidytext, refinr, 
               lubridate, glue, 
               stringr, knitr, testthat)
source("../code/get_label.R")
source("../code/clean_levels.R")
source("../code/clean_n.R")
source("../code/get_inferred.R")
source("../code/clean_pv.R")
source("../code/add_cohort.R")
source("../code/get_DA.R")
```

# Read in the data

```{r}
TSH  <- read_excel("../raw data/20191206_excel TSHV3.xlsx")
TSH
```


# Cleaning of the data

## Missing rows

Blank rows between papers in excel spreadsheet - remove. 

```{r}
TSH  <- 
  TSH %>% 
  filter(!is.na(Paper))
TSH
```


## Add an ID

```{r}
TSH$ID  <- seq_along(TSH$Paper)
```

## Paper

```{r, fig.asp = 1.5}
TSH %>% 
  count(Paper) %>% 
  ggplot(aes(fct_reorder(Paper, n), n)) + 
  geom_bar(stat = "identity") + 
  coord_flip() + 
  labs(x = NULL, y = "Number of results")
```

## Continuous versus quantile

First rename to make easier

```{r}
TSH  <- 
  TSH %>% 
  rename(CvQ = `Continuous V quantile`)
```

```{r}
TSH %>% 
  count(CvQ)
```

Use the following set of mappings:

```{r}
TSH  <- 
  TSH %>% 
  mutate(
    CvQ = str_to_lower(CvQ), 
    CvQ = case_when(
      str_detect(CvQ, "cont") ~ "continuous",
      str_detect(CvQ, "dicho") ~ "dichotomous",
      str_detect(CvQ, "quart") ~ "quartile",
      str_detect(CvQ, "performance") ~ "performance",
      TRUE ~ CvQ
    )
  )
```

```{r}
TSH %>% 
  count(CvQ)
```

## Condition

```{r}
TSH %>% 
  count(`Condition Classification`)
```

Some conditions, e.g., are present for more than one system.

```{r}
TSH %>% 
  count(`Condition Classification`, System)
```

So we will create a condition that is of the form:

- `system : condition`

```{r}
TSH  <- TSH %>% 
  mutate(
    condition = str_c(System, ":", `Condition Classification`)
  ) 
TSH %>% 
  select(System, `Condition Classification`, condition)
```

## System

```{r}
TSH %>% count(System) %>% 
  ggplot(aes(fct_reorder(System, n), n)) + 
  geom_bar(stat = 'identity') + 
  labs(x = "System")
```


## FT4_N

```{r}
table(TSH$FT4_N)
```

So some seem to have cases and controls. I will use the total as N, also not stated I will see to NA. Finally approx will be set to the value. 

```{r code=readLines("../code/clean_n.R"), eval = FALSE}
```

```{r}
clean_n(TSH$FT4_N, check = TRUE)
```

```{r}
TSH  <- 
  TSH %>%
  mutate(
  FT4_N = clean_n(FT4_N)
  )
```

```{r}
summary(TSH$FT4_N)
```


## FT4 method

```{r}
TSH %>% 
  count(FT4_method)
```

```{r}
clean_levels(TSH$FT4_method, check = TRUE)
```

```{r}
TSH  <- 
  TSH %>% 
  mutate(
    FT4_method = clean_levels(FT4_method)
  )
```


## FT4 sig

```{r}
TSH %>% count(FT4_Sig)
```

Convert `N*` to `N`. 

```{r code=readLines("../code/get_inferred.R"), eval = FALSE}
```

```{r}
TSH  <- 
  TSH %>% 
  mutate(FT4_inferred = get_inferred(FT4_Sig))
table(TSH$FT4_Sig, TSH$FT4_inferred)
TSH  <- 
  TSH %>% 
  mutate(FT4_Sig = str_replace_all(FT4_Sig, "\\*", ""))
table(TSH$FT4_Sig, TSH$FT4_inferred)
```


## FT4 PV

Function to take PV column and convert to PV, OR CI, and count CI. 

```{r code=readLines("../code/clean_pv.R"), eval = FALSE}
```

```{r}
clean_pv(TSH$FT4_PV, check = TRUE)
```

```{r}
pv  <- clean_pv(TSH$FT4_PV)
pv  <- pv %>% select(-origin_x)
pv
TSH  <- 
  TSH %>% 
  bind_cols(pv)
```

```{r}
TSH %>% 
  ggplot(aes(FT4_pv)) + geom_histogram(col = "black", fill = "orange")
```


## FT4_cov

```{r}
clean_levels(TSH$FT4_cov, check = TRUE)
```

```{r}
TSH  <- 
  TSH %>% 
  mutate(
    FT4_cov = clean_levels(FT4_cov)
  )
```

## FT4_sophistication

```{r}
TSH %>% count(FT4_Sophistication)
TSH %>% filter(is.na(FT4_Sophistication))
```

## TSH_N

```{r}
clean_n(TSH$TSH_N, check = TRUE)
TSH  <- TSH %>% 
  mutate(TSH_N = clean_n(TSH_N))
```

## TSH_method

```{r}
clean_levels(TSH$TSH_method, check = TRUE)
TSH  <- 
  TSH %>% 
  mutate(TSH_method = clean_levels(TSH_method))
```

## TSH_sig

```{r}
TSH %>% 
  count(TSH_Sig)
```

```{r}
TSH  <- 
  TSH %>% 
  mutate(TSH_inferred = get_inferred(TSH_Sig))
table(TSH$TSH_Sig, TSH$TSH_inferred)
TSH  <- 
  TSH %>% 
  mutate(TSH_Sig = str_replace_all(TSH_Sig, "\\*", ""))
table(TSH$TSH_Sig, TSH$TSH_inferred)
```


## TSH PV

```{r}
clean_pv(TSH$TSH_PV, check = TRUE)
```

```{r}
pv  <- clean_pv(TSH$TSH_PV)
pv  <- pv %>% select(-origin_x)
pv
TSH  <- 
  TSH %>% 
  bind_cols(pv)
```

```{r}
TSH %>% 
  ggplot(aes(TSH_pv)) + geom_histogram(col = "black", fill = "orange")
```

## TSH cov

```{r}
clean_levels(TSH$TSH_cov, check = TRUE)
TSH  <- 
  TSH %>% 
  mutate(
    TSH_cov = clean_levels(TSH_cov)
  )
```

## TSH_sophistication

```{r}
TSH %>% count(TSH_Sophistication)
TSH %>% filter(is.na(TSH_Sophistication))
```


## T3 N

```{r}
clean_n(TSH$T3_N, check = TRUE)
TSH  <- 
  TSH %>% 
  mutate(
    T3_N = clean_n(T3_N)
  )
```

## T3 method

```{r}
clean_levels(TSH$T3_method, check = TRUE)
TSH  <- 
  TSH %>% 
  mutate(
    T3_method = clean_levels(T3_method)
  )
```

## T3 sig

```{r}
TSH %>% 
  count(T3_Sig)
```

```{r}
TSH %>% 
  count(T3_Sig)
```

```{r}
TSH  <- 
  TSH %>% 
  mutate(T3_inferred = get_inferred(T3_Sig))
table(TSH$T3_Sig, TSH$T3_inferred)
TSH  <- 
  TSH %>% 
  mutate(T3_Sig = str_replace_all(T3_Sig, "\\*", ""))
table(TSH$T3_Sig, TSH$T3_inferred)
```


## T3 PV

```{r}
clean_pv(TSH$T3_PV, check = TRUE)
```

```{r}
pv  <- clean_pv(TSH$T3_PV)
pv  <- pv %>% select(-origin_x)
pv
TSH  <- 
  TSH %>% 
  bind_cols(pv)
```

```{r}
TSH %>% 
  ggplot(aes(T3_pv)) + geom_histogram(col = "black", fill = "orange")
```

## T3 cov

```{r}
clean_levels(TSH$T3_cov, check = TRUE)
TSH  <- 
  TSH %>% 
  mutate(
    T3_cov = clean_levels(T3_cov)
  )
```

## T3_sophistication

```{r}
TSH %>% count(T3_Sophistication)
```

## Total T3

```{r}
TSH %>% count(`T3 - Free or Total`)
```

Rename

```{r}
TSH  <- 
  TSH %>% 
  rename(
    T3_total = `T3 - Free or Total`
  )
```

```{r}
TSH %>% count(T3_total)
```

## Remove original PV columns

```{r}
TSH  <- 
  TSH %>% 
  select(-FT4_PV, -TSH_PV, -T3_PV)
```

## Add cohort column


```{r code=readLines("../code/add_cohort.R"), eval = FALSE}
```

```{r}
TSH  <- 
  TSH %>% group_by(Paper) %>% 
  mutate(cohort = add_cohort(FT4_N))
```

```{r, fig.asp=2}
TSH %>% ggplot(aes(cohort, FT4_N)) + geom_point() + 
  facet_wrap(~Paper, scales = "free", ncol = 3)
```

## Check hyper / hypo

```{r}
TSH %>% 
  ungroup() %>% 
  count(`hyper/hypo`)
```

```{r}
TSH  <- TSH %>% 
  mutate(`hyper/hypo` = str_to_title(`hyper/hypo`))
```

```{r}
TSH %>% 
  ungroup() %>% 
  count(`hyper/hypo`)
```

## Devil advocate sampling

For each paper, we select one analysis from the following treatments:

- low size, low number of covariates
- low size, large number of covariates
- large size, low number of covariates
- large size, large number of covariates

Samples first for size and then for method. May get repeated analysis, but ensures an analysis for each treatment level. 

```{r}
set.seed(2019)
TSH  <-
  TSH %>%
  group_by(Paper) %>%
  mutate(
    min_size_min_method = get_DA(
      FT4_N, FT4_cov, size = 'min', method = 'min'),
    min_size_max_method = get_DA(
      FT4_N, FT4_cov, size = 'min', method = 'max'),
    max_size_min_method = get_DA(
      FT4_N, FT4_cov, size = 'max', method = 'min'),
    max_size_max_method = get_DA(
      FT4_N, FT4_cov, size = 'max', method = 'max')
  )
```

So there should be one for each paper. 

```{r}
length(unique(TSH$Paper))
TSH %>% filter(min_size_min_method)
TSH %>% filter(min_size_max_method)
TSH %>% filter(max_size_min_method)
TSH %>% filter(max_size_max_method)
```

Sample to compare
```{r}
TSH %>% 
  filter(Paper == TSH$Paper[1], min_size_min_method) %>% 
  select(Paper, FT4_N, FT4_cov)
```
```{r}
TSH %>% 
  filter(Paper == TSH$Paper[1], min_size_max_method) %>% 
  select(Paper, FT4_N, FT4_cov)
```
```{r}
TSH %>% 
  filter(Paper == TSH$Paper[1], max_size_min_method) %>% 
  select(Paper, FT4_N, FT4_cov)
```

```{r}
TSH %>% 
  filter(Paper == TSH$Paper[1], max_size_max_method) %>% 
  select(Paper, FT4_N, FT4_cov)
```

# Save clean data

```{r}
write_rds(TSH, glue("../data/20191207-01_cleaning-data.rds"))
```
