==> analysis/02_convert_long.rmd <==
---
title: "Convert to long form"
author: "Jono Tuke"
date: "Dec 2019"
output: 
  html_document:
    theme: spacelab
    number_sections: yes
    toc_float: yes
    toc: yes
    df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE, 
  fig.width = 6, 
  fig.asp = 0.618, 
  out.width = "70%",
  fig.align = "center", 
  root.dir = '../'
)
pacman::p_load(
  tidyverse, lubridate, stringr, glue
)
```

# Read in clean data

```{r}
TSH  <- read_rds("../data/20191207-01_cleaning-data.rds")
TSH
```

# Remove grouping

```{r}
TSH  <- TSH %>% ungroup()
TSH
```

# Convert to long form

First take the the significance so row for each significance result. The key measure is made up of two parts - thanks NB for listening - this means that we can split into measure and T4 form

```{r}
TSH_long  <-
  TSH %>%
  ungroup() %>% 
  pivot_longer(
    cols = dplyr::matches("FT4_|TSH_|T3_"),
    names_to = c("Thyroid_test", ".value"),
    names_sep = "_" 
  )
TSH_long
```

# Add in extra treatment

So we will split T3 into T3_total and T3_free

```{r}
TSH_long <- 
  TSH_long %>% 
  mutate(
    Thyroid_test_4 = ifelse(
      is.na(total), 
      Thyroid_test, 
      paste0(Thyroid_test, "_",total)
    )
  )
TSH_long %>% 
  count(Thyroid_test_4)
```

# Number of covariates in the model

```{r}
TSH_long$n_cov  <- str_split(TSH_long$cov, ",") %>% map(length) %>% unlist()
table(TSH_long$n_cov)
```

Will split into 1, 2, 3, more

```{r}
TSH_long <- 
  TSH_long %>% 
  mutate(
    n_cov = case_when(
      n_cov == 1 ~ "1", 
      n_cov == 2 ~ "2", 
      n_cov == 3 ~ "3", 
      TRUE ~ "more"
    )
  )
```

# N

```{r}
TSH_long$N  <- as.numeric(TSH_long$N)
TSH_long
TSH_long %>% ggplot(aes(N)) + geom_histogram(col = "black", fill = "orange")
summary(TSH_long$N)
TSH_long  <- 
  TSH_long %>% 
  mutate(N_bin = cut_number(N, 5))
TSH_long %>% filter(is.na(N_bin))
```

```{r}
TSH_long
```




# Write data

```{r}
write_rds(TSH_long, glue("../data/20191207-02_convert_long.rds"))
```
