==> analysis/03b_ci_plots_filtered.Rmd <==
---
title: "CI plots -filtered"
author: "Jono Tuke"
date: "Dec 2019"
output: 
  html_document:
    theme: spacelab
    number_sections: yes
    toc_float: yes
    toc: yes
    df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE, 
  fig.width = 6, 
  fig.asp = 0.618, 
  out.width = "70%",
  fig.align = "center", 
  root.dir = '../'
)
pacman::p_load(
  tidyverse, lubridate, stringr, glue, lme4, lmerTest
)
source("../code/visualise_ci.R")
```

# Read in clean data

```{r}
TSH_long  <- read_rds("../data/20191207-02_convert_long.rds")
TSH_long
```

# Filter

We have columns called `include?`

```{r}
TSH_long %>% count(`Include?`)
```

I assume that the mapping is 

- 0 do not include, 
- 1 include. 

We will filter all results on include. 

```{r}
TSH_long  <- 
  TSH_long %>% 
  filter(`Include?` == 1)
```

# Visual comparison

## Overall

```{r}
TSH_long
overall_ci  <- 
  TSH_long %>% 
  filter(!is.na(Sig)) %>% 
  group_by(Thyroid_test) %>% 
  summarise(
    n_sig = sum(Sig == "Y"), 
    N = n(), 
    p = n_sig / N,
    lwr = p - 1.96 * sqrt(p * (1 - p) / N),
    upr = p + 1.96 * sqrt(p * (1 - p) / N)
  )
overall_ci
ggplot(overall_ci, aes( Thyroid_test, p, col = Thyroid_test)) + 
  geom_point() + 
  geom_errorbar(aes(ymin = lwr, ymax = upr), width = 0.1) + 
  theme(legend.position = "none") + 
  labs(y = "Proportion")
```


## System

```{r}
visualise_ci(TSH_long, System)
```


## System

```{r}
visualise_ci(TSH_long, condition)
```

## Method

```{r, fig.asp=1}
visualise_ci(TSH_long, method)
```

Too many to see a good pattern. 

## N

```{r}
visualise_ci(TSH_long, N_bin)
```

Looks like some interaction effect


## Number of covariates in model

```{r}
visualise_ci(TSH_long, n_cov)
```

Again some interaction. 


## Extra treatment

In this case, we will have each of the T3s on there own and a final combined one. 

```{r}
TSH_long %>% count(Thyroid_test_4)
```


```{r}
overall_ci  <- TSH_long %>% 
  filter(str_detect(Thyroid_test_4, "T3")) %>% 
  filter(!is.na(Sig)) %>% 
  group_by(Thyroid_test_4) %>% 
  summarise(
    n_sig = sum(Sig == "Y"), 
    N = n(), 
    p = n_sig / N,
    lwr = p - 1.96 * sqrt(p * (1 - p) / N),
    upr = p + 1.96 * sqrt(p * (1 - p) / N)
  )
overall_ci
combinded  <- tibble(
  Thyroid_test_4 = "Combined", 
  n_sig = sum(overall_ci$n_sig),
  N = sum(overall_ci$N)
) %>% 
  mutate(
    p = n_sig / N,
    lwr = p - 1.96 * sqrt(p * (1 - p) / N),
    upr = p + 1.96 * sqrt(p * (1 - p) / N)
  )
combinded
```

```{r}
ggplot(overall_ci, aes( Thyroid_test_4, p, col = Thyroid_test_4)) + 
  geom_point() + 
  geom_errorbar(aes(ymin = lwr, ymax = upr), width = 0.1) + 
  theme(legend.position = "none") + 
  geom_point(aes(Thyroid_test_4, p), data = combinded, shape = 17, size = 3) + 
  geom_errorbar(aes(ymin = lwr, ymax = upr), width = 0.1, data = combinded, linetype = "dashed", size = 1) + 
  labs(y = "Proportion")
```


## Hyper / hypo

```{r}
TSH_long
visualise_ci(TSH_long, `hyper/hypo`)
```
